+++ ocr_parser/contracts/__init__.py
@@ -3,6 +3,7 @@
 from .document import ArtifactMetadata, DocumentResult, MarkdownArtifacts, PageTask
 from .engine import EngineCapabilities, EngineContext, EnginePageResult, OCREngine
 from .events import JobEventPayload
+from .execution import EngineExecutionTrace, FallbackInfo, StageOutcome
 from .failure import FailureCategory
 from .manifest import ManifestItem
 
@@ -12,10 +13,13 @@ __all__ = [
     "EngineCapabilities",
     "EngineContext",
     "EnginePageResult",
+    "EngineExecutionTrace",
+    "FallbackInfo",
     "FailureCategory",
     "JobEventPayload",
     "ManifestItem",
     "MarkdownArtifacts",
     "OCREngine",
     "PageTask",
+    "StageOutcome",
 ]